Fixed file select dialog opening unexpectedly. Closes #64
This issue arose when the user, editing a category widget, double-clicked an item in the mobile emulator. Because the photo uploader flash object was still adjusting itself, it was exposed just long enough for the second click in the double click to catch it, thus opening the file select dialog.