Make image resizing based on original size. Closes #89
The easy way was to do width: x%. However css does not support specifying
a width based on an image's original size. Fulfilling this requirement
required javascript code that dynamically retrieves the current width
and height of the original image (luckily already present as a preview
image), and using it to manually set the width and height in pixels of
the target image.