Fixed ie7 jwysiwyg image dialog bugs. Closes #113
This commit fixes two bugs.

The first bug caused the image to reduce to nothing on dialog save. This
was because ie cannot read the size of an image with style display:none.
To work around this, in addition to adding display:block, I also added
visibility:hidden, as well as positioned the image absolutely to avoid the
image from flickering in appearance.

The second bug was ie7's lack of support for the string method trim().
Because of ie7's severe lack of debugging tools, this was, at best, fairly
difficult to track down.