Fixed page_tree page saving bug. Closes #114
The main reason this bug was happening is because of the emulator update
queue. The queue's purpose is to not update the emulator on every jwysiwyg
key press. However, in a slower browser, the user might hit the save button
before the queued update takes place, resulting in lost changes.

This is resolved by clearing the queue and updating, if the queue is not
empty.