Implemented jwysiwyg photo uploader cancelling. Closes #118
It's not actually possible to cancel an upload, but it is possible to make
it look like it has. When the user clicks cancel in the middle of an upload,
the dialog closes and the uploader is marked as "cancelled". A "cancelled"
uploader does nothing once the upload is complete.

Normally the same uploader is reused for each dialog to save performance and
memory. However, if the user cancels an upload, then clicks 'upload photo'
again when the previous upload is not done, Bad Things might happen. To get
around this, a new, temporary uploader is created if a previous uploader is
marked as cancelled and is still in the upload process.