Added font-size to _yomobi.header.sass to fix font size difference
There was a font size difference between the locomotive cms header and
the signin / signup header and I think it's because the font sizes are
relative - in em - and the baseline for the two headers was not the
same. By adding a fixed font-size, it should make the baseline the same
for both.