Updated code to make the New Text Area dialog appear
Modified files so when you click on the New Text Area button in a
Category editor, the New Text Area dialog appears. There's still a lot
of stuff to be fixed, like removing the Photo section from the New Text
Area dialog, since there is no photo associated with a Text Area,
adding an Icon, pulling up the correct editor when you edit a text area.